Little bit of refactoring (add playerclass) Rebuild input class back to using WinAPI More consistent input class implementation (working both mouse and keyboard) Deletion of unnecessary part of d3dclass from early adhoc test versions

+++ EngineDemo/timerclass.cpp
@@ -11,12 +11,12 @@ TimerClass::TimerClass()
 
 float TimerClass::TotalTime() const
 {
-	return (float)mTotalTime*mSecondsPerCount;
+	return static_cast<float>(static_cast<double>(mTotalTime)*mSecondsPerCount);
 }
 
 float TimerClass::DeltaTime() const
 {
-	return mDeltaTime;
+	return static_cast<float>(mDeltaTime);
 }
 
 void TimerClass::Reset()
